requirements.txt
graphviz==0.20.3
matplotlib==3.8.2
numpy==1.23.0
pandas==1.4.3
pycryptodome==3.20.0
scikit_learn==1.1.2
sympy==1.12.1
